littleupdate

+++ cogs/owner.py
@@ -11,12 +11,16 @@ class Owner(commands.Cog):
     @commands.command()
     @commands.has_any_role(*valid_roles)
     async def leave_server(self, ctx):
+        if ctx.author != ctx.guild.owner:
+            return
         await ctx.send(f"{ctx.author.mention} I'm leaving!!!")
         await self.bot.get_guild(int(ctx.guild.id)).leave()
     
     @commands.command()
     @commands.has_any_role(*valid_roles)
     async def kick(self, ctx, user: discord.Member, *, reason=None):
+        if ctx.author != ctx.guild.owner:
+            return
         if ctx.author.guild_permissions.kick_members:
             await user.kick(reason=reason)
             await ctx.send(f"{user.mention} has been kicked from the server. Reason: {reason}")
@@ -26,6 +30,8 @@ class Owner(commands.Cog):
     @commands.command()
     @commands.has_any_role(*valid_roles)
     async def addrole(self, ctx, role_name: str, hex_color: str = "", user: discord.Member = None):
+        if ctx.author != ctx.guild.owner:
+            return
         if user is None:
             user = ctx.author
 
@@ -54,6 +60,8 @@ class Owner(commands.Cog):
     @commands.command()
     @commands.has_any_role(*valid_roles)
     async def removerole(self, ctx, role_name: str, user: discord.Member = None):
+        if ctx.author != ctx.guild.owner:
+            return
         if user is None:
             user = ctx.author
 
@@ -71,6 +79,8 @@ class Owner(commands.Cog):
     @commands.command(aliases=["prename","pname"])
     @commands.has_any_role(*valid_roles)
     async def pnickname(self, ctx, new_nickname, user=None):
+        if ctx.author != ctx.guild.owner:
+            return
         await ctx.message.delete()
         if not user:
             target_user = ctx.author
@@ -84,6 +94,25 @@ class Owner(commands.Cog):
             await target_user.edit(nick=new_nickname)
         except discord.Forbidden:
             pass
+    
+    @commands.command(aliases=["pc"])
+    @commands.has_any_role(*valid_roles)
+    async def purgechannel(self, ctx, channel_id=None):
+        if ctx.author != ctx.guild.owner:
+            return
+        if channel_id is None:
+            channel_id = ctx.channel.id
+
+        channel = self.bot.get_channel(int(channel_id))
+        if channel is None:
+            await ctx.send("Invalid channel ID.")
+            return
+
+        try:
+            await channel.purge(limit=None)
+            await ctx.send(f"All messages deleted in {channel.mention}.")
+        except discord.Forbidden:
+            await ctx.send("I don't have permission to delete messages in that channel.")
 
 async def setup(bot):
     await bot.add_cog(Owner(bot))
